Sales tax cannot be reverse calculated
Hi Sean - I did not see anywhere to upload the images within MKDocs - so I used the example of image file load path, but no images are actually there to match it. 

Also - this should be an issue maybe rather than a commit, but I am pretty sure this is correct as I have documented it, but some-one else may need to review? as you said you were not across this issue. 

Also - to note that this is tested on 4.6 but I also tested it on the demo site of drupal and civi 4.7 and the results were the same.

Hope this helps,
Renaee.